requirements.txt
langchain
python-dotenv
openai
fitz
pdfplumber
streamlit
faiss-cpu
altair
tiktoken
langchain-community
PyPDF2
